requirements.txt
fastapi
google-generativeai
uvicorn
python-dotenv